requirements.txt
# Python 3.12.7
torch==2.5.1
transformers==4.46.3
faiss-cpu==1.9.0.post1
pymupdf==1.24.14
pytesseract==0.3.13
Pillow==11.0.0
tqdm==4.67.1
ipywidgets==8.1.5
matplotlib==3.9.2